Update index to force Python integer type

+++ src/yolo.py
@@ -229,7 +229,10 @@ class ExperimentManager:
             }
 
             self.db.add_records(
-                list(patches["class"]), vectors, metadata, list(patches.index)
+                list(patches["class"]),
+                vectors,
+                metadata,
+                list(patches.index.astype(int)),
             )
         # For subsequent calls, return the max ID value placed in dataset
         return patch_df.index[-1]
